Almost Done
Here, this is a correct solution for the problem.  I just have to complete the add() function, which I will do tomorrow.

Sorry for the ridiculous naming of the array DirtyHarry.  I just watched the Dirty Harry movie and I thought it's cool that I have this big gun ready to shoot the problem.